% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{basin_model_methods}
\alias{basin_model_methods}
\alias{print.basin_model}
\alias{plot.basin_model}
\title{Print and plot methods for basin model results}
\usage{
\method{print}{basin_model}(x, ...)

\method{plot}{basin_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"basin_model"}.}

\item{...}{For \code{plot}, further arguments passed on to
\code{\link{plot_basin_model}}. Not used by \code{print}.}
}
\value{
\code{x}, invisibly.
}
\description{
Methods for objects of class \code{"basin_model"}, the result of
\code{\link{run_basin_model}}. \code{print} gives a short summary of the
modelled burial and temperature history and \code{plot} draws the composite
figure of \code{\link{plot_basin_model}}.
}
\examples{
res <- run_basin_model(brabant)
print(res)
plot(res)
}
\seealso{
\code{\link{run_basin_model}}, \code{\link{plot_basin_model}}
}
\author{
Michiel Arts
}
